server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv

pydantic[email]
pydantic-settings

python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1

python-multipart  
openpyxl
